Fix issue with pagination links on nested pages
`next_page` and `prev_page` appear to generate absolute paths but **without** the leading `/` which means outputting them directly leads the user to a 404 page instead of the next or previous page.

Quickly fix this by just appending a leading slash when outputting the path. `Middleman::Util#normalize_path` appears to be the culprit, it's called on the blog's `page_link` option.