Consistently use date & lastmod in frontmatter for published & modified dates
We can't depend on the git info - as vercel does a shallow clone - as it doesn't look only for changes to posts, but also other unrelated changes to content within the post.

Instead find the handful of posts we have amended post-publish and explicitly set their `lastmod` datetime in the frontmatter. This is something we'll have to remember to do going forwards.